refactor(reusable-components): cleaned code
