+++ src/shared/TextInputWithLabel.jsx
@@ -1,0 +1,24 @@
+
+function TextInputWithLabel({
+    elementId,
+    label,
+    onChange,
+    ref,
+    value
+}) {
+
+    return (
+        <>
+            <label htmlFor={elementId}>{label}</label>
+            <input
+                type="text"
+                id={elementId}
+                ref={ref}
+                value={value}
+                onChange={onChange}
+            />
+        </>
+    )
+}
+
+export default TextInputWithLabel;
